drivers: staging: android: split uapi out of binder.h
Move the userspace interface of binder.h to
drivers/staging/android/uapi/binder.h.

Git-commit: 06f505a4d5719da00e76bde885792a7d5ec968f8
Git-repo: https://android.googlesource.com/kernel/common/
